+++ dfk_lib/auctions/auction.py
