+++ dfk_lib/auctions/auction_core.py
