+++ dfk_lib/auctions/hero/rent_auctions.py
