+++ dfk_lib/auctions/hero/sale_auctions.py
