+++ dfk_lib/auctions/utils/utils.py
