+++ dfk_lib/examples/hero_example.py
@@ -1,9 +1,9 @@
 import logging
 import json
 import sys
-import hero.utils.utils as hero_utils
-import hero.hero as heroes
-
+import dfk_lib.hero.utils.utils as hero_utils
+import dfk_lib.hero.hero as heroes
+from pathlib import Path
 
 if __name__ == "__main__":
     log_format = '%(asctime)s|%(name)s|%(levelname)s: %(message)s'
@@ -15,16 +15,16 @@ if __name__ == "__main__":
     rpc_server = 'https://api.harmony.one'
     logger.info("Using RPC server " + rpc_server)
 
-    with open('hero/femaleFirstName.json', 'r') as f:
-        female_first_names = hero_utils.parse_names(f.read())
+    hero_dir = Path(__file__).parent.parent / 'hero'
+    females_path = hero_dir / 'femaleFirstName.json'
+    males_path = hero_dir / 'maleFirstName.json'
+    last_names_path = hero_dir / 'lastName.json'
+
+    female_first_names = hero_utils.parse_names(females_path.read_text())
     logger.info("Female hero first name loaded")
-
-    with open('hero/maleFirstName.json', 'r') as f:
-        male_first_names = hero_utils.parse_names(f.read())
+    male_first_names = hero_utils.parse_names(males_path.read_text())
     logger.info("Male hero first name loaded")
-
-    with open('hero/lastName.json', 'r') as f:
-        last_names = hero_utils.parse_names(f.read())
+    last_names = hero_utils.parse_names(last_names_path.read_text())
     logger.info("Hero last name loaded")
 
     # transfer(1, 'private key of the owner', 'next nonce of owner account', 'receiver address', 200, rpc_server, hero_abi_json, logger)
