+++ dfk_lib/examples/wishing_well_quest_example.py
