+++ dfk_lib/perilous_journey/perilous_journey.py
