+++ dfk_lib/quest/fishing.py
