+++ dfk_lib/quest/foraging.py
