+++ dfk_lib/quest/gardening.py
