+++ dfk_lib/quest/minning.py
