+++ dfk_lib/quest/quest.py
