+++ dfk_lib/quest/quest_core.py
