+++ dfk_lib/quest/utils/utils.py
