+++ dfk_lib/quest/wishing_well.py
